backend/requirements.txt
flask
flask-cors
openai-whisper
torch
numpy
ffmpeg-python
